NET-2: moved connection package from hostnetwork to transport, fixed license
