+++ network/transport/connection/doc.go
